requirements.txt
alpha_mini_rug==0.4.1
autobahn
msgpack
numpy
pillow
pyOpenSSL
serialization
SpeechRecognition
Twisted
google-genai
python-dotenv
pyphen
py-feat